前台逻辑之电影弹幕(FLask结合Redis消息队列实现电影弹幕)

+++ app/home/views.py
@@ -1,12 +1,12 @@
 # -*- coding: utf-8 -*-
 __author__ = 'QB'
 from . import home
-from flask import render_template, redirect, url_for, flash, session, request
+from flask import render_template, redirect, url_for, flash, session, request, Response
 from app.home.forms import RegisterForm, LoginForm, UserdetailForm, PwdForm, CommentForm
 from app.models import User, Userlog, Preview, Tag, Movie, Comment, Moviecol
 from werkzeug.security import generate_password_hash
 from werkzeug.utils import secure_filename
-from app import db, app
+from app import db, app, rd
 from functools import wraps
 import urllib
 import json
@@ -15,7 +15,7 @@ import os
 from datetime import datetime
 
 # 分页个数
-PAGE_COUNT = 10
+PAGE_COUNT = 1
 
 
 # 登录装饰器
@@ -135,9 +135,10 @@ def user():
             os.chmod(app.config['DC_DIR'], 'rw')
         # 上传头像
         if form.face.data != '':  # 说明已经重新上传了头像
-            # 如果有头像，删除本地的头像。保存新的头像在本地
-            if os.path.exists(app.config['UP_DIR'] + user.face):
-                os.remove(app.config['UP_DIR'] + user.face)
+            if user.face is not None:
+                # 如果有头像，删除本地的头像。保存新的头像在本地
+                if os.path.exists(app.config['UP_DIR'] + user.face):
+                    os.remove(app.config['UP_DIR'] + user.face)
             file_face = secure_filename(form.face.data.filename)
             user.face = change_filename(file_face)
             form.face.data.save(app.config['FC_DIR'] + user.face)
@@ -243,7 +244,6 @@ def moviecol_add():
         db.session.add(moviecol)
         db.session.commit()
         data = dict(ok=1)
-    import json
     return json.dumps(data)
 
 
@@ -390,3 +390,92 @@ def play(id=None, page=None):
     db.session.add(movie)
     db.session.commit()
     return render_template('home/play.html', movie=movie, form=form, page_data=page_data)
+
+
+# 弹幕播放器
+@home.route("/video/<int:id>/<int:page>/", methods=["GET", "POST"])
+def video(id=None, page=None):
+    movie = Movie.query.join(Tag).filter(
+        Tag.id == Movie.tag_id,
+        Movie.id == int(id)
+    ).first_or_404()
+
+    if page is None:
+        page = 1
+    page_data = Comment.query.join(
+        Movie
+    ).join(
+        User
+    ).filter(
+        Movie.id == movie.id,
+        User.id == Comment.user_id
+    ).order_by(
+        Comment.addtime.desc()
+    ).paginate(page=page, per_page=PAGE_COUNT)
+
+    form = CommentForm()
+    # 判断是否已经登录
+    if "user" in session and form.validate_on_submit():
+        data = form.data
+        comment = Comment(
+            content=data['content'],
+            movie_id=movie.id,
+            user_id=session['user_id'],
+        )
+        db.session.add(comment)
+        db.session.commit()
+        movie.commentnum = movie.commentnum + 1
+        db.session.add(movie)
+        db.session.commit()
+        flash("添加评论成功！", "ok")
+        return redirect(url_for("home.video", id=movie.id, page=1))
+    movie.playnum = movie.playnum + 1
+    db.session.add(movie)
+    db.session.commit()
+    return render_template("home/video.html", movie=movie, form=form, page_data=page_data)
+
+
+# 弹幕消息处理
+@home.route("/tm/", methods=["GET", "POST"])
+def tm():
+    if request.method == "GET":
+        # 获取弹幕消息队列
+        id = request.args.get('id')
+        # 存放在redis队列中的键值
+        key = "movie" + str(id)
+        if rd.llen(key):
+            msgs = rd.lrange(key, 0, 2999)
+            res = {
+                "code": 1,
+                "danmaku": [json.loads(v) for v in msgs]
+            }
+        else:
+            res = {
+                "code": 1,
+                "danmaku": []
+            }
+        resp = json.dumps(res)
+    if request.method == "POST":
+        # 添加弹幕
+        data = json.loads(request.get_data())
+        msg = {
+            "__v": 0,
+            "author": data["author"],
+            "time": data["time"],
+            "text": data["text"],
+            "color": data["color"],
+            "type": data['type'],
+            "ip": request.remote_addr,
+            "_id": datetime.now().strftime("%Y%m%d%H%M%S") + uuid.uuid4().hex,
+            "player": [
+                data["player"]
+            ]
+        }
+        res = {
+            "code": 1,
+            "data": msg
+        }
+        resp = json.dumps(res)
+        # 将添加的弹幕推入redis的队列中
+        rd.lpush("movie" + str(data["player"]), json.dumps(msg))
+    return Response(resp, mimetype='application/json')
